# docs/openapi-spec.yaml
openapi: 3.0.3
info:
  title: Golf Gamez API
  description: |
    API for tracking golf game scores and side bets for weekend warriors.

    ## Features
    - Anonymous game creation with shareable links
    - Real-time score tracking for up to 4 players
    - Two side bets: Best Nine and Putt Putt Poker
    - Live leaderboards and spectator mode
    - Diamond Run golf course integration

    ## Authentication
    Uses token-based access with two token types:
    - **Share Token**: Full read/write access for players
    - **Spectator Token**: Read-only access for viewing

    ## Rate Limiting
    - 10 game creations per hour per IP
    - 100 score updates per minute per IP
    - 1000 API calls per hour per IP
  version: 1.0.0
  contact:
    name: Golf Gamez Support
    url: https://github.com/golf-gamez/api
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.golfgamez.com/v1
    description: Production server
  - url: https://staging-api.golfgamez.com/v1
    description: Staging server
  - url: http://localhost:8080/v1
    description: Development server
security:
  - GameToken: []
  - SpectatorToken: []
  - {}
paths:
...
  /games/{gameId}/side-bets/best-nine:
    get:
      summary: Get Best Nine standings
      description: Retrieve current Best Nine side bet standings
      operationId: getBestNineStandings
      parameters:
        - $ref: '#/components/parameters/GameId'
      responses:
        '200':
          description: Best Nine standings retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BestNineStandings'
  /games/{gameId}/side-bets/putt-putt-poker:
    get:
      summary: Get Putt Putt Poker status
      description: Retrieve current Putt Putt Poker side bet status
      operationId: getPuttPuttPokerStatus
      parameters:
        - $ref: '#/components/parameters/GameId'
      responses:
        '200':
          description: Putt Putt Poker status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PuttPuttPokerStatus'
  /games/{gameId}/side-bets/putt-putt-poker/deal:
    post:
      summary: Deal final poker cards
      description: Deal final cards and determine poker hand winner
      operationId: dealPokerCards
      parameters:
        - $ref: '#/components/parameters/GameId'
      responses:
        '201':
          description: Cards dealt successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PokerDealResult'
...
components:
  securitySchemes:
    GameToken:
      type: http
      scheme: bearer
      bearerFormat: gt_abc123def456ghi789
      description: Game share token for full read/write access
    SpectatorToken:
      type: http
      scheme: bearer
      bearerFormat: st_abc123def456ghi789
      description: Spectator token for read-only access
  parameters:
    GameId:
      name: gameId
      in: path
      required: true
      description: Unique game identifier or share token
      schema:
        type: string
        pattern: '^(game_[a-zA-Z0-9]{15}|gt_[a-zA-Z0-9]{20})$'
        example: "game_abc123def456"
...
  schemas:
...
    GameStatus:
      type: string
      enum: ["setup", "in_progress", "completed", "abandoned"]
      description: |
        Game status values:
        - setup: Game created but not started
        - in_progress: Game is active
        - completed: All holes finished
        - abandoned: Game ended early
...
    BestNineStandings:
      type: object
      properties:
        bet_type:
          type: string
          enum: ["best_nine"]
        status:
          $ref: '#/components/schemas/GameStatus'
        handicap_enabled:
          type: boolean
        standings:
          type: array
          items:
            $ref: '#/components/schemas/BestNineResult'
        winner:
          type: object
          properties:
            player_id:
              type: string
            score:
              type: string
            margin:
              type: string
    BestNineResult:
      type: object
      properties:
        player:
          $ref: '#/components/schemas/PlayerSummary'
        best_nine_score:
          type: string
          example: "+2"
        holes_completed:
          type: integer
        best_holes:
          type: array
          items:
            type: integer
        worst_holes:
          type: array
          items:
            type: integer
        raw_best_nine:
          type: string
        handicap_adjustment:
          type: string
        final_score:
          type: string
        position:
          type: integer
    PuttPuttPokerStatus:
      type: object
      properties:
        bet_type:
          type: string
          enum: ["putt_putt_poker"]
        status:
          $ref: '#/components/schemas/GameStatus'
        current_hole:
          type: integer
        players:
          type: array
          items:
            $ref: '#/components/schemas/PuttPuttPokerResult'
        pot_info:
          type: object
          properties:
            base_bet:
              type: number
              format: float
            penalty_additions:
              type: number
              format: float
            total_pot:
              type: number
              format: float
    PuttPuttPokerResult:
      type: object
      properties:
        player:
          $ref: '#/components/schemas/PlayerSummary'
        total_cards:
          type: integer
        starting_cards:
          type: integer
        cards_earned:
          type: integer
        penalties:
          type: integer
        putting_stats:
          $ref: '#/components/schemas/PuttingStats'
        position:
          type: integer
    PuttingStats:
      type: object
      properties:
        one_putts:
          type: integer
        hole_in_ones:
          type: integer
        three_putts:
          type: integer
        average_putts:
          type: number
          format: float
    PokerDealResult:
      type: object
      properties:
        deal_timestamp:
          type: string
          format: date-time
        random_seed:
          type: string
        players:
          type: array
          items:
            $ref: '#/components/schemas/PokerHand'
        winner:
          type: object
          properties:
            player_id:
              type: string
            hand_type:
              type: string
            winning_cards:
              type: array
              items:
                type: string
        pot_distribution:
          type: object
          properties:
            total_pot:
              type: number
              format: float
            winner_take:
              type: number
              format: float
    PokerHand:
      type: object
      properties:
        player:
          $ref: '#/components/schemas/PlayerSummary'
        total_cards_earned:
          type: integer
        dealt_cards:
          type: array
          items:
            type: string
            example: "AS"
        best_hand:
          type: object
          properties:
            cards:
              type: array
              items:
                type: string
            hand_type:
              type: string
              enum: ["royal_flush", "straight_flush", "four_of_a_kind", "full_house", "flush", "straight", "three_of_a_kind", "two_pair", "pair", "high_card"]
            hand_rank:
              type: integer
              minimum: 1
              maximum: 10
            description:
              type: string
        position:
          type: integer
    PlayerSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        handicap:
          type: number
...